finish up cableready refactor
